Revert "feat: Implement database monitoring and health check services with Op…"
This reverts commit ae5758c9eff7984cf23faa5e4974e0bc71ed30e0.
