+++ frontend/src/HomePage/AppCard.jsx
@@ -157,11 +157,6 @@ export default function AppCard({
             )}
             disabled={app?.current_version_id === null || app?.is_maintenance_on}
             onClick={() => {
-              // Store start time for Launch button load tracking (using Date.now for cross-page timing)
-              const loadStartTime = Date.now();
-              localStorage.setItem(`app_launch_load_start_${app.id}`, loadStartTime.toString());
-              localStorage.setItem(`app_launch_mode_${app.id}`, 'launch');
-              
               if (app?.current_version_id) {
                 window.open(
                   urlJoin(window.public_config?.TOOLJET_HOST, getSubpath() ?? '', `/applications/${app.slug}`)
@@ -277,11 +272,6 @@ export default function AppCard({
                       slug: isValidSlug(app.slug) ? app.slug : app.id,
                     })}
                     onClick={() => {
-                      // Store start time for Edit button load tracking (using Date.now for cross-page timing)
-                      const loadStartTime = Date.now();
-                      localStorage.setItem(`app_edit_load_start_${app.id}`, loadStartTime.toString());
-                      localStorage.setItem(`app_edit_mode_${app.id}`, 'edit');
-                      
                       posthogHelper.captureEvent('click_edit_button_on_card', {
                         workspace_id:
                           authenticationService?.currentUserValue?.organization_id ||
