+++ server/src/helpers/database.helper.ts
@@ -1,7 +1,6 @@
 import { DataSource, EntityManager } from 'typeorm';
 import { updateTimestampForAppVersion } from './utils.helper';
 import { createLogger } from './bootstrap.helper';
-import { databaseMonitoring } from '../otel/monitoring/database-monitoring';
 
 let CONNECTION_INSTANCE: DataSource;
 const getConnectionInstance = (): DataSource => {
@@ -15,12 +14,6 @@ export function setConnectionInstance(dataSource: DataSource) {
   const logger = createLogger('setConnectionInstance');
   logger.log('Database connection initialized');
   CONNECTION_INSTANCE = dataSource;
-  
-  // Initialize database monitoring if OTEL is enabled
-  if (process.env.ENABLE_OTEL === 'true') {
-    databaseMonitoring.setDataSource(dataSource);
-    logger.log('Database monitoring initialized with OTEL');
-  }
 }
 
 export async function dbTransactionWrap(operation: (...args) => any, manager?: EntityManager): Promise<any> {
