+++ server/src/modules/ai/service.ts
@@ -1,5 +1,4 @@
 import { Injectable } from '@nestjs/common';
-import { InstrumentService } from '../../otel/business/service-instrumentation';
 import { IAiService } from './interfaces/IService';
 
 @Injectable()
@@ -18,10 +17,6 @@ export class AiService implements IAiService {
     throw new Error('Method not implemented.');
   }
 
-  @InstrumentService('AiService', { 
-    attributes: { 'operation.type': 'message', 'ai.type': 'user_message' },
-    tags: { 'business_operation': 'ai_user_message' }
-  })
   async sendUserMessage(body, response, organizationId) {
     throw new Error('Method not implemented.');
   }
