+++ server/src/modules/workflows/services/workflow-executions.service.ts
@@ -1,5 +1,4 @@
 import { Injectable } from '@nestjs/common';
-import { InstrumentService } from '../../../otel/business/service-instrumentation';
 import { AppVersion } from 'src/entities/app_version.entity';
 import { WorkflowExecution } from 'src/entities/workflow_execution.entity';
 import { User } from 'src/entities/user.entity';
@@ -17,10 +16,6 @@ export class WorkflowExecutionsService implements IWorkflowExecutionsService {
     throw new Error('Method not implemented.');
   }
 
-  @InstrumentService('WorkflowExecutionsService', { 
-    attributes: { 'operation.type': 'execute' },
-    tags: { 'business_operation': 'workflow_execution' }
-  })
   async execute(
     workflowExecution: WorkflowExecution,
     params: Record<string, any>,
