--- server/src/otel/analysis/index.ts
@@ -1,10 +1,0 @@
-/**
- * Performance Analysis & Benchmarking
- *
- * Exports for benchmarking framework and performance analysis tools.
- */
-
-// Release performance comparison and benchmarking
-export * from './benchmarking-framework';
-
-// Note: Common types are exported from the main otel/index.ts
