--- server/src/otel/business/index.ts
@@ -1,16 +1,0 @@
-/**
- * Business Logic & Application Tracking
- *
- * Exports for business metrics, application tracing, and service instrumentation.
- */
-
-// Business KPIs and user metrics
-export * from './business-metrics';
-
-// Application lifecycle tracing
-export * from './application-tracing';
-
-// Service layer instrumentation
-export * from './service-instrumentation';
-
-// Note: Common types are exported from the main otel/index.ts
