--- server/src/otel/core/tracing.ts
@@ -1,601 +1,0 @@
-import { CompositePropagator, W3CTraceContextPropagator, W3CBaggagePropagator } from '@opentelemetry/core';
-import { trackApiCall } from '../business/business-metrics';
-import { trace, context, Span, DiagConsoleLogger, DiagLogLevel, diag, metrics } from '@opentelemetry/api';
-
-// Import application-level tracing utilities
-import {
-  traceAppLifecycleOperation,
-  traceQueryExecution,
-  traceDataSourceConnection,
-  traceUserJourney,
-  addUserActivityEvent,
-  createCorrelationContext,
-  linkOperation,
-  recordBusinessError,
-  recordErrorRecovery,
-  getCurrentTraceContext,
-  createChildSpan
-} from '../business/application-tracing';
-import { ApplicationContext, BusinessOperationContext } from '../types';
-import { BatchSpanProcessor } from '@opentelemetry/sdk-trace-node';
-import { NodeSDK } from '@opentelemetry/sdk-node';
-import * as process from 'process';
-import { resourceFromAttributes } from '@opentelemetry/resources';
-import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-http';
-import { ExpressInstrumentation } from '@opentelemetry/instrumentation-express';
-import { HttpInstrumentation } from '@opentelemetry/instrumentation-http';
-import { NestInstrumentation } from '@opentelemetry/instrumentation-nestjs-core';
-import { PgInstrumentation } from '@opentelemetry/instrumentation-pg';
-import { PinoInstrumentation } from '@opentelemetry/instrumentation-pino';
-import { RuntimeNodeInstrumentation } from '@opentelemetry/instrumentation-runtime-node';
-import { OTLPMetricExporter } from '@opentelemetry/exporter-metrics-otlp-http';
-import { PeriodicExportingMetricReader } from '@opentelemetry/sdk-metrics';
-import {
-  ATTR_SERVICE_NAME,
-  ATTR_SERVICE_VERSION,
-  SEMRESATTRS_DEPLOYMENT_ENVIRONMENT,
-  SEMATTRS_DB_STATEMENT,
-  SEMATTRS_DB_OPERATION,
-  SEMATTRS_DB_NAME,
-  SEMATTRS_DB_SQL_TABLE,
-} from '@opentelemetry/semantic-conventions';
-
-const OTEL_EXPORTER_OTLP_TRACES = process.env.OTEL_EXPORTER_OTLP_TRACES || 'http://localhost:4318/v1/traces';
-const OTEL_EXPORTER_OTLP_METRICS = process.env.OTEL_EXPORTER_OTLP_METRICS || 'http://localhost:4318/v1/metrics';
-
-// Database monitoring configuration
-const DB_SLOW_QUERY_THRESHOLD = parseInt(process.env.DB_SLOW_QUERY_THRESHOLD || '1000'); // ms
-const DB_ENABLE_QUERY_ANALYSIS = process.env.DB_ENABLE_QUERY_ANALYSIS !== 'false';
-
-// Set this up to see debug logs
-if (process.env.OTEL_LOG_LEVEL === 'debug') {
-  diag.setLogger(new DiagConsoleLogger(), DiagLogLevel.DEBUG);
-}
-
-// Define the trace exporter
-const traceExporter = new OTLPTraceExporter({
-  url: OTEL_EXPORTER_OTLP_TRACES,
-  ...(!!process.env.OTEL_HEADER ? { headers: { Authorization: process.env.OTEL_HEADER } } : {} ),
-});
-
-// Define the metric exporter
-const metricExporter = new OTLPMetricExporter({
-  url: OTEL_EXPORTER_OTLP_METRICS,
-  ...(!!process.env.OTEL_HEADER ? { headers: { Authorization: process.env.OTEL_HEADER } } : {} ),
-});
-
-// Define the log exporter
-// TODO:
-// Add logs exporter when stable support for JS is available. Track here:
-// https://github.com/open-telemetry/opentelemetry-js
-
-const resource = resourceFromAttributes({
-  [ATTR_SERVICE_NAME]: process.env.SERVICE_NAME || 'tooljet',
-  [ATTR_SERVICE_VERSION]: globalThis.TOOLJET_VERSION || process.env.SERVICE_VERSION || 'unknown',
-  [SEMRESATTRS_DEPLOYMENT_ENVIRONMENT]: process.env.NODE_ENV || 'development',
-});
-
-const sanitizeObject = (obj: any) => {
-  const sanitized: any = {};
-  for (const [key, value] of Object.entries(obj)) {
-    if (typeof value === 'object' && value !== null) {
-      sanitized[key] = sanitizeObject(value);
-    } else if (typeof value !== 'function') {
-      sanitized[key] = value;
-    }
-  }
-  return sanitized;
-};
-
-// Database query analysis utilities
-const extractTableNames = (query: any): string[] => {
-  if (!query || typeof query !== 'string') {
-    return [];
-  }
-
-  const normalizedQuery = query.toLowerCase().replace(/\s+/g, ' ').trim();
-  const tables: Set<string> = new Set();
-
-  // Match common SQL patterns
-  const patterns = [
-    /from\s+([a-zA-Z_][a-zA-Z0-9_]*)/g,
-    /join\s+([a-zA-Z_][a-zA-Z0-9_]*)/g,
-    /update\s+([a-zA-Z_][a-zA-Z0-9_]*)/g,
-    /insert\s+into\s+([a-zA-Z_][a-zA-Z0-9_]*)/g,
-    /delete\s+from\s+([a-zA-Z_][a-zA-Z0-9_]*)/g,
-  ];
-
-  patterns.forEach(pattern => {
-    let match;
-    while ((match = pattern.exec(normalizedQuery)) !== null) {
-      if (match[1] && !match[1].includes('(')) {
-        tables.add(match[1]);
-      }
-    }
-  });
-
-  return Array.from(tables);
-};
-
-const getQueryOperation = (query: any): string => {
-  if (!query || typeof query !== 'string') {
-    return 'OTHER';
-  }
-  const normalizedQuery = query.toLowerCase().trim();
-  if (normalizedQuery.startsWith('select')) return 'SELECT';
-  if (normalizedQuery.startsWith('insert')) return 'INSERT';
-  if (normalizedQuery.startsWith('update')) return 'UPDATE';
-  if (normalizedQuery.startsWith('delete')) return 'DELETE';
-  if (normalizedQuery.startsWith('create')) return 'CREATE';
-  if (normalizedQuery.startsWith('drop')) return 'DROP';
-  if (normalizedQuery.startsWith('alter')) return 'ALTER';
-  return 'OTHER';
-};
-
-// Sanitize SQL query following OpenTelemetry conventions
-// Replace literals with ? placeholders but preserve query structure
-const sanitizeQueryForMetrics = (query: any): string => {
-  if (!query) {
-    return 'unknown_query';
-  }
-
-  try {
-    // Convert query to string safely
-    let queryString = '';
-    if (typeof query === 'string') {
-      queryString = query;
-    } else if (query && typeof query === 'object') {
-      // Handle query objects (common with some PostgreSQL drivers)
-      if ('text' in query && typeof query.text === 'string') {
-        queryString = query.text;
-      } else if ('query' in query && typeof query.query === 'string') {
-        queryString = query.query;
-      } else {
-        queryString = JSON.stringify(query);
-      }
-    } else {
-      queryString = String(query);
-    }
-
-    // OpenTelemetry sanitization: replace literals with ? but preserve structure
-    let sanitized = queryString
-      .trim()
-      .replace(/\s+/g, ' ') // Normalize whitespace
-      .replace(/\$\d+/g, '?') // Replace parameterized values ($1, $2, etc.) with ?
-      .replace(/'[^']*'/g, '?') // Replace string literals 'value' with ?
-      .replace(/"[^"]*"/g, '?') // Replace quoted identifiers "table" with ?
-      .replace(/\b\d+\.?\d*\b/g, '?') // Replace numeric literals with ?
-      .replace(/\b[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}\b/gi, '?') // Replace UUIDs
-      .replace(/\b[a-f0-9]{32}\b/gi, '?'); // Replace MD5 hashes
-
-    // Limit length for performance (OpenTelemetry recommends 255 chars max)
-    if (sanitized.length > 255) {
-      sanitized = sanitized.substring(0, 252) + '...';
-    }
-
-    return sanitized;
-  } catch (error) {
-    return 'parse_error';
-  }
-};
-
-const isSlowQuery = (duration: number): boolean => {
-  return duration >= DB_SLOW_QUERY_THRESHOLD;
-};
-
-// Initialize custom metrics
-let dbQueryDurationHistogram: any;
-let dbSlowQueryCounter: any;
-let dbConnectionPoolGauge: any;
-let dbQueryCounter: any;
-
-// Store request information for response processing
-const spanRequestMap = new Map<string, any>();
-
-// Lazy initialization to prevent issues during migrations
-let sdk: NodeSDK | null = null;
-
-const createSDK = () => {
-  if (!sdk) {
-    sdk = new NodeSDK({
-      resource: resource,
-      traceExporter: traceExporter,
-      spanProcessors: [new BatchSpanProcessor(traceExporter)], // Fixed deprecated option
-      metricReader: new PeriodicExportingMetricReader({
-        exporter: metricExporter,
-        exportIntervalMillis: 10000, // Export every 10 seconds for debugging
-      }),
-      textMapPropagator: new CompositePropagator({
-        propagators: [new W3CTraceContextPropagator(), new W3CBaggagePropagator()],
-      }),
-  instrumentations: [
-    new RuntimeNodeInstrumentation({ monitoringPrecision: 5000 }),
-    new HttpInstrumentation({
-      ignoreIncomingRequestHook: (request: any) => request.url === '/api/health',
-    }),
-    new ExpressInstrumentation({
-      requestHook: (span: Span, { request }) => {
-        const path = request.route?.path || request.path || '';
-        if (path.startsWith('/api/') && path !== '/api/health') {
-          span.updateName(`${request.method} ${path}`);
-          span.setAttribute('http.route', path);
-          span.setAttribute('http.method', request.method);
-
-          if (request.params && Object.keys(request.params).length > 0) {
-            span.setAttribute('http.params', JSON.stringify(sanitizeObject(request.params)));
-          }
-
-          if (request.query && Object.keys(request.query).length > 0) {
-            span.setAttribute('http.query', JSON.stringify(sanitizeObject(request.query)));
-          }
-
-          if (request.body && Object.keys(request.body).length > 0) {
-            span.setAttribute('http.body', JSON.stringify(sanitizeObject(request.body)));
-          }
-        }
-      },
-    }),
-    new NestInstrumentation(),
-    new PgInstrumentation({ 
-      enhancedDatabaseReporting: true,
-      requestHook: (span: Span, requestInfo: any) => {
-        if (DB_ENABLE_QUERY_ANALYSIS && requestInfo.query) {
-          const query = requestInfo.query;
-          const startTime = Date.now();
-
-          // Convert query to string safely
-          let queryString = '';
-          let queryLength = 0;
-
-          if (typeof query === 'string') {
-            queryString = query;
-            queryLength = query.length;
-          } else if (query && typeof query === 'object') {
-            // Handle query objects (common with some PostgreSQL drivers)
-            if (query.text) {
-              queryString = query.text;
-              queryLength = query.text.length;
-            } else if (query.query) {
-              queryString = query.query;
-              queryLength = query.query.length;
-            } else {
-              queryString = JSON.stringify(query);
-              queryLength = queryString.length;
-            }
-          } else {
-            queryString = String(query || '');
-            queryLength = queryString.length;
-          }
-
-          // Store request info for response hook using span context
-          const spanContext = span.spanContext();
-          spanRequestMap.set(spanContext.spanId, {
-            startTime,
-            query: queryString,
-            connectionParameters: requestInfo.connectionParameters
-          });
-
-          // Extract query operation and table names
-          const operation = getQueryOperation(queryString);
-          const tables = extractTableNames(queryString);
-
-          // Add custom attributes
-          span.setAttribute(SEMATTRS_DB_OPERATION, operation);
-          span.setAttribute('db.query.tables', tables.join(','));
-          span.setAttribute('db.query.length', queryLength);
-          span.setAttribute('db.query.start_time', startTime);
-
-          // Add query parameters if available
-          if (query && typeof query === 'object' && query.values && Array.isArray(query.values)) {
-            span.setAttribute('db.query.params_count', query.values.length);
-            // Don't log actual parameter values for security reasons
-          }
-
-          // Sanitize and add query statement (limit length for performance)
-          const sanitizedQuery = queryString.length > 1000
-            ? queryString.substring(0, 1000) + '...[truncated]'
-            : queryString;
-          span.setAttribute(SEMATTRS_DB_STATEMENT, sanitizedQuery);
-
-          // Add connection info
-          if (requestInfo.connectionParameters) {
-            span.setAttribute('db.connection.host', requestInfo.connectionParameters.host || 'unknown');
-            span.setAttribute('db.connection.port', requestInfo.connectionParameters.port || 5432);
-            span.setAttribute('db.connection.database', requestInfo.connectionParameters.database || 'unknown');
-          }
-        }
-      },
-      responseHook: (span: Span, responseInfo: any) => {
-        if (DB_ENABLE_QUERY_ANALYSIS) {
-          const spanContext = span.spanContext();
-          const storedInfo = spanRequestMap.get(spanContext.spanId);
-          
-          if (storedInfo) {
-            const endTime = Date.now();
-            const startTime = storedInfo.startTime;
-            const duration = endTime - startTime;
-            
-            // Add response metadata
-            span.setAttribute('db.query.duration_ms', duration);
-            span.setAttribute('db.query.end_time', endTime);
-
-            // Add result information
-            if (responseInfo.data) {
-              if (responseInfo.data.rowCount !== undefined) {
-                span.setAttribute('db.query.rows_affected', responseInfo.data.rowCount);
-              }
-              if (responseInfo.data.rows && Array.isArray(responseInfo.data.rows)) {
-                span.setAttribute('db.query.rows_returned', responseInfo.data.rows.length);
-              }
-            }
-
-            // Add error information if query failed
-            if (responseInfo.error) {
-              span.setAttribute('db.query.error', true);
-              span.setAttribute('db.query.error_message', responseInfo.error.message || 'Unknown error');
-              span.setAttribute('db.query.error_code', responseInfo.error.code || 'UNKNOWN');
-            } else {
-              span.setAttribute('db.query.error', false);
-            }
-            
-            // Mark slow queries
-            if (isSlowQuery(duration)) {
-              span.setAttribute('db.query.is_slow', true);
-              span.setAttribute('db.query.slow_threshold_ms', DB_SLOW_QUERY_THRESHOLD);
-              
-              // Record slow query metric
-              if (dbSlowQueryCounter) {
-                const operation = getQueryOperation(storedInfo.query);
-                dbSlowQueryCounter.add(1, {
-                  'db.operation.name': operation.toLowerCase(),
-                  'db.namespace': storedInfo.connectionParameters?.database || 'unknown',
-                  'db.system': 'postgresql'
-                });
-                console.log('[ToolJet Backend] Slow query detected:', {
-                  duration: duration,
-                  operation: operation.toLowerCase(),
-                  database: storedInfo.connectionParameters?.database
-                });
-              }
-            }
-            
-            // Record query duration histogram (convert ms to seconds)
-            // Initialize metrics on first use if not already initialized
-            if (!dbQueryDurationHistogram) {
-              console.log('[ToolJet Backend] Initializing database metrics lazily');
-              const meter = metrics.getMeter('tooljet-database', '1.0.0');
-              dbQueryDurationHistogram = meter.createHistogram('db.client.operation.duration', {
-                description: 'Duration of database client operations.',
-                unit: 's'
-                // Note: Histogram boundaries are typically set at the collector/exporter level
-              });
-              dbSlowQueryCounter = meter.createCounter('db.client.operation.slow_total', {
-                description: 'Total number of slow database operations.',
-              });
-              dbQueryCounter = meter.createCounter('db.client.operation.count', {
-                description: 'Total number of database operations.',
-              });
-            }
-
-            if (dbQueryDurationHistogram) {
-              const operation = getQueryOperation(storedInfo.query);
-              const tables = extractTableNames(storedInfo.query);
-
-              // Create a sanitized query identifier
-              const queryPattern = sanitizeQueryForMetrics(storedInfo.query);
-
-              // Build attributes following OpenTelemetry conventions
-              // NOTE: Avoiding db.query.text as it creates high cardinality that crashes Prometheus exporter
-              const attributes = {
-                'db.system.name': 'postgresql',
-                'db.operation.name': operation.toLowerCase(),
-                'db.namespace': storedInfo.connectionParameters?.database || 'unknown',
-                'db.collection.name': tables.length > 0 ? tables[0] : 'unknown',
-                'db.query.summary': `${operation.toLowerCase()} ${tables.length > 0 ? tables[0] : 'table'}` // Low cardinality summary
-              };
-
-              dbQueryDurationHistogram.record(duration / 1000, attributes);
-            }
-            
-            // Record query counter
-            if (dbQueryCounter) {
-              const operation = getQueryOperation(storedInfo.query);
-              const tables = extractTableNames(storedInfo.query); // Need to extract tables here too
-              const queryPattern = sanitizeQueryForMetrics(storedInfo.query);
-
-              // Build counter attributes following OpenTelemetry conventions
-              // NOTE: Avoiding db.query.text as it creates high cardinality that crashes Prometheus exporter
-              const counterAttributes = {
-                'db.system.name': 'postgresql',
-                'db.operation.name': operation.toLowerCase(),
-                'db.namespace': storedInfo.connectionParameters?.database || 'unknown',
-                'db.collection.name': tables.length > 0 ? tables[0] : 'unknown',
-                'db.query.summary': `${operation.toLowerCase()} ${tables.length > 0 ? tables[0] : 'table'}` // Low cardinality summary
-              };
-
-              dbQueryCounter.add(1, counterAttributes);
-            }
-            
-            // Clean up stored info to prevent memory leaks
-            spanRequestMap.delete(spanContext.spanId);
-          }
-        }
-      }
-    }),
-    new PinoInstrumentation(),
-  ],
-    });
-  }
-  return sdk;
-};
-
-// Helper function to determine activity type from route and method
-const getActivityTypeFromRoute = (route: string, method: string): string => {
-  const methodLower = method.toLowerCase();
-
-  // App-related activities
-  if (route.includes('/api/apps')) {
-    if (methodLower === 'post') return 'app_create';
-    if (methodLower === 'put') return 'app_update';
-    if (methodLower === 'delete') return 'app_delete';
-    if (route.includes('validate-private-app-access')) return 'app_access';
-    return 'app_view';
-  }
-
-  // Query-related activities
-  if (route.includes('/api/data-queries')) {
-    if (methodLower === 'post') return 'query_execute';
-    if (methodLower === 'put') return 'query_update';
-    return 'query_view';
-  }
-
-  // Data source activities
-  if (route.includes('/api/data-sources')) {
-    return 'datasource_access';
-  }
-
-  // Authentication activities
-  if (route.includes('/api/authorize') || route.includes('/api/session')) {
-    return 'authentication';
-  }
-
-  // Organization management
-  if (route.includes('/api/organization')) {
-    return 'organization_management';
-  }
-
-  // License and limits
-  if (route.includes('/api/license')) {
-    return 'license_check';
-  }
-
-  // AI features
-  if (route.includes('/api/ai')) {
-    return 'ai_feature_usage';
-  }
-
-  // Default for other activities
-  return 'general_api_usage';
-};
-
-// Custom Express middleware for tracing and metrics
-export const otelMiddleware = (req: any, res: any, next: () => void, ...args: any[]) => {
-  const span = trace.getSpan(context.active());
-  const route = req.route?.path || req.path || 'unknown_route';
-  const method = req.method || 'UNKNOWN_METHOD';
-
-  if (span && route.startsWith('/api/') && route !== '/api/health') {
-    const startTime = Date.now();
-    
-    span.updateName(`${method} ${route}`);
-    span.setAttribute('http.route', route);
-    span.setAttribute('http.method', method);
-
-    const originalJson = res.json;
-    res.json = function (body: any) {
-      const statusCode = res.statusCode;
-      const duration = (Date.now() - startTime) / 1000; // Convert to seconds
-
-      span.setAttribute('http.status_code', statusCode);
-      
-      // Extract organization ID from request context
-      const organizationId = req.user?.organizationId ||
-                            req.headers?.['tj-workspace-id'] ||
-                            req.session?.current_organization_id ||
-                            undefined;
-
-      // Track API call business metrics
-      trackApiCall(route, method, statusCode, duration, organizationId);
-
-      // Track user activity for authenticated requests
-      if (req.user?.id && organizationId) {
-        const { updateUserActivity } = require('../business/business-metrics');
-        const activityType = getActivityTypeFromRoute(route, method);
-
-        // Update user's last activity time
-        updateUserActivity(req.user.id, organizationId, activityType);
-      }
-      
-      // eslint-disable-next-line prefer-rest-params
-      return originalJson.apply(this, arguments);
-    };
-  }
-
-  next();
-};
-
-process.on('SIGTERM', () => {
-  if (sdk) {
-    sdk
-      .shutdown()
-      .then(() => console.log('OpenTelemetry instrumentation shutdown successfully'))
-      .catch((err) => console.log('Error shutting down OpenTelemetry instrumentation', err))
-      .finally(() => process.exit(0));
-  } else {
-    process.exit(0);
-  }
-});
-
-export const startOpenTelemetry = async (): Promise<void> => {
-  // Don't initialize OTEL during database migrations
-  const isMigration = process.argv.some(arg =>
-    arg.includes('migration') ||
-    arg.includes('typeorm') ||
-    arg.includes('cli.js')
-  );
-
-  if (isMigration) {
-    console.log('[ToolJet Backend] Skipping OpenTelemetry initialization during migration');
-    return;
-  }
-
-  try {
-    const sdkInstance = createSDK();
-    await sdkInstance.start();
-    
-    // Note: Database metrics are now lazily initialized in the PostgreSQL instrumentation hooks
-    // to avoid race conditions between SDK startup and first database queries
-
-    // Initialize database monitoring metrics
-    const { databaseMonitoring } = await import('../monitoring/database-monitoring');
-    const meter = metrics.getMeter('tooljet-database', '1.0.0');
-    databaseMonitoring.initializeMetrics(meter);
-    
-    // Initialize service layer metrics
-    const { initializeServiceMetrics } = await import('../business/service-instrumentation');
-    initializeServiceMetrics();
-    
-    // Initialize business metrics
-    const { initializeBusinessMetrics, setupResourceMetricCallbacks } = await import('../business/business-metrics');
-    initializeBusinessMetrics();
-    setupResourceMetricCallbacks();
-
-    // Initialize comprehensive API monitoring
-    const { initializeComprehensiveApiMonitoring } = await import('../monitoring/comprehensive-api-middleware');
-    await initializeComprehensiveApiMonitoring();
-
-    console.log('[ToolJet Backend] OpenTelemetry instrumentation initialized with enhanced database, service, business, and comprehensive API monitoring');
-  } catch (error) {
-    console.error('Error initializing OpenTelemetry instrumentation', error);
-    throw error;
-  }
-};
-
-// Export application-level tracing functions for use throughout the application
-export {
-  traceAppLifecycleOperation,
-  traceQueryExecution,
-  traceDataSourceConnection,
-  traceUserJourney,
-  addUserActivityEvent,
-  createCorrelationContext,
-  linkOperation,
-  recordBusinessError,
-  recordErrorRecovery,
-  getCurrentTraceContext,
-  createChildSpan,
-  ApplicationContext,
-  BusinessOperationContext
-};
-
-export default sdk;
