--- server/src/otel/monitoring/index.ts
@@ -1,16 +1,0 @@
-/**
- * Performance Monitoring Modules
- *
- * Exports for API performance metrics, database monitoring, and middleware instrumentation.
- */
-
-// API performance tracking
-export * from './api-performance-metrics';
-
-// Database monitoring (consolidated enhanced + basic functionality)
-export * from './database-monitoring';
-
-// API middleware instrumentation
-export * from './comprehensive-api-middleware';
-
-// Note: Common types are exported from the main otel/index.ts
