+++ server/src/otel/tracing.ts
@@ -1,0 +1,154 @@
+import { CompositePropagator, W3CTraceContextPropagator, W3CBaggagePropagator } from '@opentelemetry/core';
+import { trace, context, Span, DiagConsoleLogger, DiagLogLevel, diag } from '@opentelemetry/api';
+import { BatchSpanProcessor } from '@opentelemetry/sdk-trace-node';
+import { NodeSDK } from '@opentelemetry/sdk-node';
+import * as process from 'process';
+import { resourceFromAttributes } from '@opentelemetry/resources';
+import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-http';
+import { ExpressInstrumentation } from '@opentelemetry/instrumentation-express';
+import { HttpInstrumentation } from '@opentelemetry/instrumentation-http';
+import { NestInstrumentation } from '@opentelemetry/instrumentation-nestjs-core';
+import { PgInstrumentation } from '@opentelemetry/instrumentation-pg';
+import { PinoInstrumentation } from '@opentelemetry/instrumentation-pino';
+import { RuntimeNodeInstrumentation } from '@opentelemetry/instrumentation-runtime-node';
+import { OTLPMetricExporter } from '@opentelemetry/exporter-metrics-otlp-http';
+import { PeriodicExportingMetricReader } from '@opentelemetry/sdk-metrics';
+import {
+  ATTR_SERVICE_NAME,
+  ATTR_SERVICE_VERSION,
+  SEMRESATTRS_DEPLOYMENT_ENVIRONMENT,
+} from '@opentelemetry/semantic-conventions';
+
+const OTEL_EXPORTER_OTLP_TRACES = process.env.OTEL_EXPORTER_OTLP_TRACES || 'http://localhost:4318/v1/traces';
+const OTEL_EXPORTER_OTLP_METRICS = process.env.OTEL_EXPORTER_OTLP_METRICS || 'http://localhost:4318/v1/metrics';
+
+// Set this up to see debug logs
+if (process.env.OTEL_LOG_LEVEL === 'debug') {
+  diag.setLogger(new DiagConsoleLogger(), DiagLogLevel.DEBUG);
+}
+
+// Define the trace exporter
+const traceExporter = new OTLPTraceExporter({
+  url: OTEL_EXPORTER_OTLP_TRACES,
+  ...(!!process.env.OTEL_HEADER ? { headers: { Authorization: process.env.OTEL_HEADER } } : {} ),
+});
+
+// Define the metric exporter
+const metricExporter = new OTLPMetricExporter({
+  url: OTEL_EXPORTER_OTLP_METRICS,
+  ...(!!process.env.OTEL_HEADER ? { headers: { Authorization: process.env.OTEL_HEADER } } : {} ),
+});
+
+// Define the log exporter
+// TODO:
+// Add logs exporter when stable support for JS is available. Track here:
+// https://github.com/open-telemetry/opentelemetry-js
+
+const resource = resourceFromAttributes({
+  [ATTR_SERVICE_NAME]: process.env.SERVICE_NAME || 'tooljet',
+  [ATTR_SERVICE_VERSION]: globalThis.TOOLJET_VERSION || process.env.SERVICE_VERSION || 'unknown',
+  [SEMRESATTRS_DEPLOYMENT_ENVIRONMENT]: process.env.NODE_ENV || 'development',
+});
+
+const sanitizeObject = (obj: any) => {
+  const sanitized: any = {};
+  for (const [key, value] of Object.entries(obj)) {
+    if (typeof value === 'object' && value !== null) {
+      sanitized[key] = sanitizeObject(value);
+    } else if (typeof value !== 'function') {
+      sanitized[key] = value;
+    }
+  }
+  return sanitized;
+};
+
+export const sdk = new NodeSDK({
+  resource: resource,
+  traceExporter: traceExporter,
+  spanProcessor: new BatchSpanProcessor(traceExporter),
+  metricReader: new PeriodicExportingMetricReader({
+    exporter: metricExporter,
+  }),
+  textMapPropagator: new CompositePropagator({
+    propagators: [new W3CTraceContextPropagator(), new W3CBaggagePropagator()],
+  }),
+  instrumentations: [
+    new RuntimeNodeInstrumentation({ monitoringPrecision: 5000 }),
+    new HttpInstrumentation({
+      ignoreIncomingRequestHook: (request: any) => request.url === '/api/health',
+    }),
+    new ExpressInstrumentation({
+      requestHook: (span: Span, { request }) => {
+        const path = request.route?.path || request.path || '';
+        if (path.startsWith('/api/') && path !== '/api/health') {
+          span.updateName(`${request.method} ${path}`);
+          span.setAttribute('http.route', path);
+          span.setAttribute('http.method', request.method);
+
+          if (request.params && Object.keys(request.params).length > 0) {
+            span.setAttribute('http.params', JSON.stringify(sanitizeObject(request.params)));
+          }
+
+          if (request.query && Object.keys(request.query).length > 0) {
+            span.setAttribute('http.query', JSON.stringify(sanitizeObject(request.query)));
+          }
+
+          if (request.body && Object.keys(request.body).length > 0) {
+            span.setAttribute('http.body', JSON.stringify(sanitizeObject(request.body)));
+          }
+        }
+      },
+    }),
+    new NestInstrumentation(),
+    new PgInstrumentation({ enhancedDatabaseReporting: true }),
+    new PinoInstrumentation(),
+  ],
+});
+
+// Custom Express middleware for tracing and metrics
+export const otelMiddleware = (req: any, res: any, next: () => void, ...args: any[]) => {
+  const span = trace.getSpan(context.active());
+  const route = req.route?.path || req.path || 'unknown_route';
+  const method = req.method || 'UNKNOWN_METHOD';
+
+  if (span && route.startsWith('/api/') && route !== '/api/health') {
+    span.updateName(`${method} ${route}`);
+    span.setAttribute('http.route', route);
+    span.setAttribute('http.method', method);
+
+    const originalJson = res.json;
+    res.json = function (body: any) {
+      const statusCode = res.statusCode;
+
+      span.setAttribute('http.status_code', statusCode);
+      // eslint-disable-next-line prefer-rest-params
+      return originalJson.apply(this, arguments);
+    };
+  }
+
+  next();
+};
+
+process.on('SIGTERM', () => {
+  if (sdk) {
+    sdk
+      .shutdown()
+      .then(() => console.log('OpenTelemetry instrumentation shutdown successfully'))
+      .catch((err) => console.log('Error shutting down OpenTelemetry instrumentation', err))
+      .finally(() => process.exit(0));
+  } else {
+    process.exit(0);
+  }
+});
+
+export const startOpenTelemetry = async (): Promise<void> => {
+  try {
+    await sdk.start();
+    console.log('OpenTelemetry instrumentation initializated');
+  } catch (error) {
+    console.error('Error initializing OpenTelemetry instrumentation', error);
+    throw error;
+  }
+};
+
+export default sdk;
